Pull out backlog to its own component
